move new post button to layout
