+++ packages/next-common/components/layout/DiscussionsLayout.jsx
@@ -1,11 +1,15 @@
 import { useChainSettings } from "next-common/context/chain";
 import ListLayout from "./ListLayout";
+import PrimaryButton from "next-common/components/buttons/primaryButton";
+import { SystemPlus } from "@osn/icons/subsquare";
+import { useRouter } from "next/router";
 
 /**
  * @param {import("./ListLayout").ListLayoutProps} props
  */
 export default function DiscussionsLayout(props) {
   const chainSettings = useChainSettings();
+  const router = useRouter();
 
   const tabs = [
     {
@@ -21,6 +25,19 @@ export default function DiscussionsLayout(props) {
   return (
     <ListLayout
       tabs={chainSettings.hasPolkassemblyDiscussions && tabs}
+      summaryFooter={
+        <div className="flex justify-end">
+          <PrimaryButton
+            small
+            icon={
+              <SystemPlus className="w-4 h-4 [&_path]:fill-textPrimaryContrast" />
+            }
+            onClick={() => router.push("/posts/create")}
+          >
+            New Post
+          </PrimaryButton>
+        </div>
+      }
       {...props}
     />
   );
